Move GetTranscriptRegionsForSites into pipeline_utils and setup v2
